fixed annoying bug and fixed almost all tests

+++ littlechef/app/src/main/java/comp3350/littlechef/persistence/DataAccessObject.java
@@ -70,18 +70,18 @@ public class DataAccessObject implements DataAccess
 
         try
         {
-            String cmdString = "shutdown compact";
-            resultSet = statement.executeQuery(cmdString);
+            if(connection == null)
+            {
+                return false;
+            }
+            cmd = "shutdown compact";
+            resultSet = statement.executeQuery(cmd);
             connection.close();
             success = true;
         }
-        catch (SQLException se)
-        {
-            se.printStackTrace();
-        }
         catch (Exception e)
         {
-            e.printStackTrace();
+            processSQLError(e);
         }
 
         System.out.println(dbType + " type's database closed.");
@@ -95,6 +95,11 @@ public class DataAccessObject implements DataAccess
 
         String values;
         result = null;
+
+        if(connection == null)
+        {
+            return "connection is not open.";
+        }
 
         try
         {
@@ -139,7 +144,7 @@ public class DataAccessObject implements DataAccess
                         + "', '" + instructions.get(i)[1].replace("'","''")
                         + "', " + recipe.getId()
                         + "";
-                cmd = "INSERT INTO INSTRUCTIONS " + "(INSTRUCTION, SUBINSTRUCTION, RECIPEID)" + "VALUES(" + values + ")";
+                cmd = "INSERT INTO INSTRUCTIONS " + "(INSTRUCTION, SUBINSTRUCTION, RECIPEID)" + "VALUES(" + values + ")"+"\n";
                 updateCount = statement.executeUpdate(cmd);
                 result = checkWarning(statement, updateCount);
             }
@@ -151,7 +156,7 @@ public class DataAccessObject implements DataAccess
                         + ", '" + ingredients.get(i).getUnitType()
                         + "', '" + ingredients.get(i).getMeasurement().toString()
                         + "'";
-                cmd = "INSERT INTO INGREDIENTS " + "(RECIPEID, NAME, AMOUNT, UNITTYPE, UNIT)" + " VALUES (" + values + ")";
+                cmd += "\nINSERT INTO INGREDIENTS " + "(RECIPEID, NAME, AMOUNT, UNITTYPE, UNIT)" + " VALUES (" + values + ")"+"\n";
                 updateCount = statement.executeUpdate(cmd);
                 result = checkWarning(statement, updateCount);
             }
@@ -176,7 +181,7 @@ public class DataAccessObject implements DataAccess
                 String values = list.get(i)
                         + ", " + recipe.getId()
                         + "";
-                cmd = "INSERT INTO " + table + "(RATING, RECIPEID)" + " VALUES (" + values + ")";
+                cmd = "INSERT INTO " + table + "(RATING, RECIPEID)" + " VALUES (" + values + ")"+"\n";
                 updateCount = statement.executeUpdate(cmd);
                 result = checkWarning(statement, updateCount);
             }
@@ -195,11 +200,16 @@ public class DataAccessObject implements DataAccess
         String where = "WHERE RECIPEID=" + recipe.getId();
         result = null;
 
+        if(connection == null)
+        {
+            return "connection is not open.";
+        }
+
         try
         {
             table = "RECIPES";
             values = "NAME='" + recipe.getName().replace("'","''") + "'";
-            cmd = "UPDATE " + table + " SET " + values + " " + where;
+            cmd = "UPDATE " + table + " SET " + values + " " + where+"\n";
             updateCount = statement.executeUpdate(cmd);
             result = checkWarning(statement, updateCount);
 
@@ -221,7 +231,7 @@ public class DataAccessObject implements DataAccess
         {
             for (int i = 0; i < tables.length; i++)
             {
-                cmd = "DELETE FROM " + tables[i] + " WHERE RECIPEID=" + RECIPEID;
+                cmd = "DELETE FROM " + tables[i] + " WHERE RECIPEID=" + RECIPEID+"\n";
                 updateCount = statement.executeUpdate(cmd);
                 result = checkWarning(statement, updateCount);
             }
@@ -241,6 +251,11 @@ public class DataAccessObject implements DataAccess
         myID = EOF;
         myName = EOF;
         result = null;
+
+        if(connection == null)
+        {
+            return "connection is not open.";
+        }
 
         try
         {
@@ -307,16 +322,18 @@ public class DataAccessObject implements DataAccess
         Recipe recipe;
         String myID;
 
-        recipes = new ArrayList<Recipe>();
+        recipes = null;
         try
         {
-            cmd = "SELECT * FROM INGREDIENTS WHERE RECIPEID=" + newRecipe.getId();
+            cmd = "SELECT * FROM RECIPES WHERE RECIPEID=" + newRecipe.getId();
             resultSet = statement.executeQuery(cmd);
             // ResultSetMetaData md2 = rs3.getMetaData();
+            recipes = new ArrayList<Recipe>();
             while (resultSet.next())
             {
                 myID = resultSet.getString("RecipeID");
                 recipe = new Recipe(Integer.parseInt(myID));
+                recipe.setName(resultSet.getString("Name"));
                 cmd2 = "SELECT * FROM INGREDIENTS WHERE RECIPEID=" + myID;
                 resultSet2 = statement.executeQuery(cmd2);
                 while(resultSet2.next())
@@ -364,10 +381,16 @@ public class DataAccessObject implements DataAccess
     {
         int recipeId;
         result = null;
+
+        if(connection == null)
+        {
+            return "connection is not open.";
+        }
+
         try
         {
             recipeId = recipe.getId();
-            cmd = "DELETE FROM RECIPES WHERE RECIPEID ='"+recipeId+"'";
+            cmd = "DELETE FROM RECIPES WHERE RECIPEID ="+recipeId+"\n";
             updateCount = statement.executeUpdate(cmd);
             System.out.println("Deleted successfully.");
             result = checkWarning(statement, updateCount);
@@ -383,6 +406,12 @@ public class DataAccessObject implements DataAccess
     {
 
         result = null;
+
+        if(connection == null)
+        {
+            return "connection is not open.";
+        }
+
         try
         {
             cmd = "drop table if exists INGREDIENTS;\n" +
@@ -392,7 +421,6 @@ public class DataAccessObject implements DataAccess
                     "drop table if exists COOKINGTIMES;\n" +
                     "drop table if exists RECIPES;\n";
             connection.createStatement().executeUpdate(cmd);
-            result = checkWarning(statement, updateCount);
 
             cmd = "CREATE MEMORY TABLE RECIPES(RECIPEID INTEGER GENERATED BY DEFAULT AS IDENTITY(START WITH 0) NOT NULL PRIMARY KEY,NAME VARCHAR(64))\n" +
             "CREATE MEMORY TABLE INGREDIENTS(INGREDIENTID INTEGER GENERATED BY DEFAULT AS IDENTITY(START WITH 0) NOT NULL PRIMARY KEY,RECIPEID INTEGER NOT NULL,NAME VARCHAR(64),AMOUNT DOUBLE,UNITTYPE VARCHAR(64),UNIT VARCHAR(64),CONSTRAINT CONSTRAINT1 FOREIGN KEY(RECIPEID) REFERENCES RECIPES(RECIPEID) ON DELETE CASCADE)\n" +
@@ -401,7 +429,6 @@ public class DataAccessObject implements DataAccess
             "CREATE MEMORY TABLE TASTERATINGS(TASTERATINGID INTEGER GENERATED BY DEFAULT AS IDENTITY(START WITH 0) NOT NULL PRIMARY KEY,RATING DOUBLE,RECIPEID INTEGER NOT NULL,CONSTRAINT CONSTRAINT4 FOREIGN KEY(RECIPEID) REFERENCES RECIPES(RECIPEID) ON DELETE CASCADE)\n" +
             "CREATE MEMORY TABLE COOKINGTIMES(COOKINGTIMEID INTEGER GENERATED BY DEFAULT AS IDENTITY(START WITH 0) NOT NULL PRIMARY KEY,RATING INTEGER,RECIPEID INTEGER NOT NULL,CONSTRAINT CONSTRAINT5 FOREIGN KEY(RECIPEID) REFERENCES RECIPES(RECIPEID) ON DELETE CASCADE)\n";
             connection.createStatement().executeUpdate(cmd);
-            result = checkWarning(statement, updateCount);
             System.out.println("DATABASE RESET SUCCESSFULLY.");
         }
         catch (Exception e)
